base traj class
