+++ src/navigation/gate.py
@@ -17,25 +17,9 @@ DRIVE_FWD_THRESH = 0.95
 
 @dataclass
 class GateTrajectory:
-    # Coordinates of the trajectory
-    coordinates: np.ndarray
     # Associated Post IDs for the gate
     post1_id: int
     post2_id: int
-    # Currently tracked coordinate index along trajectory
-    cur_pt: int = 0
-    
-
-    def get_cur_pt(self) -> np.ndarray:
-        return self.coordinates[self.cur_pt]
-
-    def increment_point(self) -> bool:
-        """
-        Increments the tracked point in the trajectory, returns true if
-        the trajectory is finished
-        """
-        self.cur_pt += 1
-        return self.cur_pt >= len(self.coordinates)
 
     @classmethod
     def gate_trajectory(cls, approach_distance: float, post1_id: int, post2_id: int) -> GateTrajectory:
